backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
mediapipe
numpy
websockets
python-multipart
python-dotenv
google-generativeai
ultralytics
browser-use
langchain-google-genai
playwright